Allow changing values in combat debugger
