+++ src/main/resources/legend/game/debugger/combat_debugger.fxml
@@ -9,88 +9,90 @@
 <?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
 
-<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" stylesheets="@script_debugger.css" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="legend.game.debugger.CombatDebugger">
-   <children>
-      <HBox prefHeight="402.0" prefWidth="600.0" stylesheets="@combat_debugger.css">
-         <children>
-            <VBox prefHeight="346.0" prefWidth="182.0">
-               <children>
-                  <Label text="Battle Objects" />
-                  <ListView fx:id="bobjList" prefHeight="398.0" prefWidth="182.0" />
-               </children>
+<VBox maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" stylesheets="@script_debugger.css" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1" fx:controller="legend.game.debugger.CombatDebuggerController">
+   <HBox prefHeight="402.0" prefWidth="600.0" stylesheets="@combat_debugger.css">
+      <VBox.margin>
+         <Insets/>
+      </VBox.margin>
+      <VBox prefHeight="346.0" prefWidth="182.0">
+         <HBox.margin>
+            <Insets bottom="8.0" left="8.0" right="8.0" top="8.0"/>
+         </HBox.margin>
+         <Label text="Battle Objects"/>
+         <ListView fx:id="bobjList" prefHeight="398.0" prefWidth="182.0"/>
+      </VBox>
+      <VBox prefHeight="401.0" prefWidth="186.0">
+         <padding>
+            <Insets bottom="8.0" left="8.0" right="8.0" top="8.0"/>
+         </padding>
+         <Label text="Script"/>
+         <Button fx:id="scriptIndex" mnemonicParsing="false" onAction="#openScriptDebugger" prefHeight="26.0" prefWidth="127.0" text="Script"/>
+         <Separator prefWidth="200.0">
+            <opaqueInsets>
+               <Insets/>
+            </opaqueInsets>
+            <VBox.margin>
+               <Insets bottom="4.0" top="8.0"/>
+            </VBox.margin>
+         </Separator>
+         <HBox alignment="CENTER_LEFT" prefHeight="32.0" prefWidth="170.0">
+            <Label text="LVL">
                <HBox.margin>
-                  <Insets bottom="8.0" left="8.0" right="8.0" top="8.0" />
+                  <Insets right="4.0"/>
                </HBox.margin>
-            </VBox>
-            <VBox prefHeight="401.0" prefWidth="186.0">
-               <padding>
-                  <Insets bottom="8.0" left="8.0" right="8.0" top="8.0" />
-               </padding>
-               <children>
-                  <Label text="Script" />
-                  <Button fx:id="scriptIndex" mnemonicParsing="false" onAction="#openScriptDebugger" prefHeight="26.0" prefWidth="127.0" text="Script" />
-                  <Separator prefWidth="200.0">
-                     <opaqueInsets>
-                        <Insets />
-                     </opaqueInsets>
-                     <VBox.margin>
-                        <Insets bottom="4.0" top="8.0" />
-                     </VBox.margin>
-                  </Separator>
-                  <HBox alignment="CENTER_LEFT" prefHeight="32.0" prefWidth="170.0">
-                     <children>
-                        <Label text="LVL">
-                           <HBox.margin>
-                              <Insets right="4.0" />
-                           </HBox.margin>
-                        </Label>
-                        <Spinner fx:id="level" prefHeight="26.0" prefWidth="55.0" />
-                        <Label text="DLVL">
-                           <HBox.margin>
-                              <Insets left="4.0" right="4.0" />
-                           </HBox.margin>
-                        </Label>
-                        <Spinner fx:id="dlevel" prefHeight="26.0" prefWidth="49.0" />
-                     </children>
-                  </HBox>
-                  <HBox alignment="CENTER_LEFT" prefHeight="32.0" prefWidth="170.0">
-                     <children>
-                        <Label text="HP">
-                           <HBox.margin>
-                              <Insets right="4.0" />
-                           </HBox.margin>
-                        </Label>
-                        <Spinner fx:id="hp" prefHeight="26.0" prefWidth="65.0" />
-                        <Label text="/">
-                           <HBox.margin>
-                              <Insets left="4.0" right="4.0" />
-                           </HBox.margin>
-                        </Label>
-                        <Spinner fx:id="maxHp" prefHeight="26.0" prefWidth="65.0" />
-                     </children>
-                  </HBox>
-                  <HBox alignment="CENTER_LEFT" prefHeight="32.0" prefWidth="170.0">
-                     <children>
-                        <Label text="MP">
-                           <HBox.margin>
-                              <Insets right="4.0" />
-                           </HBox.margin>
-                        </Label>
-                        <Spinner fx:id="mp" prefHeight="26.0" prefWidth="65.0" />
-                        <Label text="/">
-                           <HBox.margin>
-                              <Insets left="4.0" right="4.0" />
-                           </HBox.margin>
-                        </Label>
-                        <Spinner fx:id="maxMp" prefHeight="26.0" prefWidth="65.0" />
-                     </children>
-                  </HBox>
-               </children>
-            </VBox>
-         </children>
-         <VBox.margin>
-            <Insets />
-         </VBox.margin>
-      </HBox>
-   </children>
+            </Label>
+            <Spinner fx:id="level" editable="true" prefHeight="26.0" prefWidth="55.0"/>
+            <Label text="DLVL">
+               <HBox.margin>
+                  <Insets left="4.0" right="4.0"/>
+               </HBox.margin>
+            </Label>
+            <Spinner fx:id="dlevel" editable="true" prefHeight="26.0" prefWidth="49.0"/>
+         </HBox>
+         <HBox alignment="CENTER_LEFT" prefHeight="32.0" prefWidth="170.0">
+            <Label text="HP">
+               <HBox.margin>
+                  <Insets right="4.0"/>
+               </HBox.margin>
+            </Label>
+            <Spinner fx:id="hp" editable="true" prefHeight="26.0" prefWidth="65.0"/>
+            <Label text="/">
+               <HBox.margin>
+                  <Insets left="4.0" right="4.0"/>
+               </HBox.margin>
+            </Label>
+            <Spinner fx:id="maxHp" editable="true" prefHeight="26.0" prefWidth="65.0"/>
+         </HBox>
+         <HBox alignment="CENTER_LEFT" prefHeight="32.0" prefWidth="170.0">
+            <VBox.margin>
+               <Insets/>
+            </VBox.margin>
+            <Label text="MP">
+               <HBox.margin>
+                  <Insets right="4.0"/>
+               </HBox.margin>
+            </Label>
+            <Spinner fx:id="mp" editable="true" prefHeight="26.0" prefWidth="65.0"/>
+            <Label text="/">
+               <HBox.margin>
+                  <Insets left="4.0" right="4.0"/>
+               </HBox.margin>
+            </Label>
+            <Spinner fx:id="maxMp" editable="true" prefHeight="26.0" prefWidth="65.0"/>
+         </HBox>
+         <Separator prefWidth="200.0">
+            <VBox.margin>
+               <Insets bottom="4.0" top="8.0"/>
+            </VBox.margin>
+         </Separator>
+         <HBox prefHeight="100.0" prefWidth="200.0">
+            <Button fx:id="refreshStats" mnemonicParsing="false" onAction="#refreshStats" text="Refresh">
+               <HBox.margin>
+                  <Insets right="4.0"/>
+               </HBox.margin>
+            </Button>
+            <Button fx:id="updateStats" mnemonicParsing="false" onAction="#updateStats" text="Update"/>
+         </HBox>
+      </VBox>
+   </HBox>
 </VBox>
